Enable drawer toggling and move to MonitorTable.tsx
