+++ packages/manager/src/features/Managed/Monitors/MonitorTable.tsx
@@ -1,3 +1,4 @@
+import { FormikBag } from 'formik';
 import { withSnackbar, WithSnackbarProps } from 'notistack';
 import * as React from 'react';
 import { compose } from 'recompose';
@@ -23,7 +24,9 @@ import TableSortCell from 'src/components/TableSortCell';
 import withManagedServices, {
   DispatchProps
 } from 'src/containers/managedServices.container';
+import { ManagedServicePayload } from 'src/services/managed';
 
+import MonitorDrawer from '../MonitorDrawer';
 import MonitorDialog from './MonitorDialog';
 import MonitorTableContent from './MonitorTableContent';
 
@@ -41,6 +44,8 @@ interface Props {
   loading: boolean;
   error?: Linode.ApiFieldError[];
 }
+
+export type FormikProps = FormikBag<CombinedProps, ManagedServicePayload>;
 
 export type CombinedProps = Props &
   WithStyles<ClassNames> &
@@ -65,6 +70,8 @@ export const MonitorTable: React.FC<CombinedProps> = props => {
   );
   const [selectedLabel, setLabel] = React.useState<string>('');
   const [isDeleting, setDeleting] = React.useState<boolean>(false);
+
+  const [drawerOpen, setDrawerOpen] = React.useState<boolean>(false);
 
   const handleOpenDialog = (id: number, label: string) => {
     setDeleteError(undefined);
@@ -92,6 +99,22 @@ export const MonitorTable: React.FC<CombinedProps> = props => {
       });
   };
 
+  const submitMonitorForm = (
+    values: ManagedServicePayload,
+    { setSubmitting, setErrors, setStatus }: FormikProps
+  ) => {
+    const { createServiceMonitor } = props;
+    createServiceMonitor({ ...values, timeout: +values.timeout })
+      .then(response => {
+        setSubmitting(false);
+        setDrawerOpen(false);
+      })
+      .catch(e => {
+        setErrors(e);
+        setSubmitting(false);
+      });
+  };
+
   return (
     <>
       <DocumentTitleSegment segment="Service Monitors" />
@@ -105,7 +128,10 @@ export const MonitorTable: React.FC<CombinedProps> = props => {
         <Grid item>
           <Grid container alignItems="flex-end">
             <Grid item className="pt0">
-              <AddNewLink onClick={() => null} label="Add a Monitor" disabled />
+              <AddNewLink
+                onClick={() => setDrawerOpen(true)}
+                label="Add a Monitor"
+              />
             </Grid>
           </Grid>
         </Grid>
@@ -187,6 +213,12 @@ export const MonitorTable: React.FC<CombinedProps> = props => {
         open={dialogOpen}
         error={deleteError}
         loading={isDeleting}
+      />
+      <MonitorDrawer
+        open={drawerOpen}
+        onClose={() => setDrawerOpen(false)}
+        onSubmit={submitMonitorForm}
+        mode="create"
       />
     </>
   );
